refactor(ai): extract Windows batch spawn helper (#11775)
